+++ MusicWatch/Classes/MatchTableDataSource.m
@@ -24,7 +24,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 - (int)numberOfRowsInTableView:(NSTableView *)tableView
 {
-	return matchAlbumArray.count;
+	return (int)matchAlbumArray.count;
 }
 
 - (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(int)row
